requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
fonttools==4.56.0
gunicorn==23.0.0
idna==3.10
jwt==1.3.1
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.5
pillow==11.1.0
platformshconfig==2.4.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
XlsxWriter==3.2.2
